requirements.txt
scrapli
scrapli[genie]
rich
pyyaml